+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ##### Enhancements
 
+* Add Junit reporter.
+  [Matthew Ellis](https://github.com/matthewellis)
+
 * LeadingWhitespaceRule is now auto correctable.  
   [masters3d](https://github.com/masters3d)